forensics/hark_on_the_wire/requirements.txt
pyshark
sqlcipher3-binary
